Rewritten login to work for both local/google.
data are stored in a shared object (scope.user) with google additional data at user.google

Now its more clear where the anyplace vs google tokens & ids are stored.

only /api/auth requests are intercepted for addin the  token,
which is now added to the headers as it should.

Google sign in is more clear. It uses the google token to verify the anyplace account, and then it
returns anyplace specific account.